Local repository modificado: eliminado cliente repo y user y tarjetas separado en diferentes repos
