--- src/main/java/org/example/client/repository/ClientsRepository.java
@@ -1,15 +1,0 @@
-package org.example.client.repository;
-
-import org.example.models.Cliente;
-
-import java.util.List;
-import java.util.UUID;
-
-public interface ClientsRepository {
-    List<Cliente> findAllClientes();
-    Cliente findClientById(UUID id);
-    List<Cliente> findClientByName(String name);
-    Cliente saveClient(Cliente client);
-    Boolean deleteCientById(UUID id);
-    Boolean deleteAllClients();
-}
